Get the sum of the relative position of the temperature optimum of each species. Relative to the mean of the environmental temperatures experienced by the community.

Get the sum of the relative position of the temperature optimum of each species. Relative to the mean of the environmental temperatures experienced by the community.

Usage

get_community_sum_rel_b_opt(temperatures, expt)

Arguments

temperatures

Connection to the database containing temperature data

expt

Dataset containing experiment information

Value

A dataset containing the sum of the relative position of the temperature optimum of each species. Relative to the mean of the environmental temperatures experienced by the community.

Examples

NULL
#> NULL